fix(app): keep desktop_backend.rs under the split-gate baseline
Move events --json companion invocation into desktop_events.rs so the frozen shell stays at 6152 lines.
