+++ .githooks/pre-commit-whitelist.ps1
@@ -208,6 +208,7 @@ $whitelistPatterns = @(
     'winsmux-app/src-tauri/scripts/*.ps1',
     'winsmux-app/src-tauri/src/control_pipe.rs',
     'winsmux-app/src-tauri/src/desktop_session_restore.rs',
+    'winsmux-app/src-tauri/src/desktop_events.rs',
     'winsmux-app/src-tauri/src/remote_debug_gate.rs',
     'winsmux-core/router/**',
     'winsmux-core/**',
